make schema more sane using graphql-tools
