+++ .babelrc
@@ -5,12 +5,12 @@
         ["es2015", { "modules": false }],
         "stage-0",
         "react"
-      ],
+      ]
     }
   },
   "presets": [
     "es2015",
     "stage-0",
     "react"
-  ],
+  ]
 }
