+++ package.json
@@ -73,6 +73,7 @@
     "express-graphql": "^0.6.7",
     "fuse.js": "^3.0.5",
     "graphql": "^0.10.5",
+    "graphql-tools": "^1.1.0",
     "gulp": "gulpjs/gulp#4.0",
     "helmet": "^3.8.1",
     "normalizr": "^3.2.3",
